[Build] Bump MediaCreation to 1.0.0-alpha.9
Part of OpenAssetIO/OpenAssetIO#1311. Deprecated aliases have been
removed from the latest OpenAssetIO, but the version of MediaCreation
that was pinned in `pyproject.toml` still used them.

So bump the version of MediaCreation to the latest.

Use a minimum, rather than fixed, version for dependencies. This
provides a useful canary. The project should be tested against each new
upstream (pending) release.

Signed-off-by: David Feltell <david.feltell@foundry.com>